Add additional factory methods to circumvent adverse method precedence
Previously, calling `factory.document(block)` (or a similar method)
actually selected `MarkdomFactory#documenty(Iterable<MarkdomBlock>)` if
the block happens to be a block parent, skipping the intended block and
erroneously adding its children to the document instead. This happened
because the method taking an `Iterable` takes precedence over the
varargs method with the same name.