Relax list block, content parent and block parent interfaces
The `getBlocks` method from `MarkdomBlockParent` (for example) now
returns a simple list of `MarkdomBlock`s which allows a seamless
interaction with that list. It is the block parents responsibility to
properly adopt and handle any block that is added to this list (or
refuse to adopt it with an appropriate exception). Block parents created
by the `BasicMarkdomFactory` can, currently, only adopt blocks that have
been created by the same kind of factory (i.e. `ManagedMarkdomBlock`s).
This could be enhanced in the future. A factory could maintain a set of
adapters that map the methods of `ManagedMarkdomBlock` to the
corresponding methods of another Markdom block implementation.